Change: (-archive-wget-options) Ignore SVG files
Too many sites seem to include large SVG webfonts, in addition to the
WOFF and EOT ones, and many seem to throw all of them into a single
page to try to cover all the bases.  These several-hundred KB are
typically 10-100x the size of a page's readable text and add up over
time.  Users who want to preserve the appearance of a page may want to
archive them (and might choose to use archive.today rather than wget,
anyway), but users who are mainly concerned about the readable text
content are probably not.  I certainly don't want a 5-10 KB web page
archive file to be bloated to 500 KB, 1 MB, or more, just for some web
fonts, which are probably also duplicated in page after page.  Of
course, SVG files are not necessarily fonts, but in most cases, they
will be, and users can always preview the archive to ensure it's what
they wanted.